requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-ratelimit==4.1.0
drf-yasg==1.21.10
django-axes==7.0.2
celery==5.4.0
daphne==4.2.1
redis==5.2.1
python-decouple==3.8
psycopg2-binary==2.9.10
requests==2.32.3
pillow==11.1.0
channels==4.1.0
channels-redis==4.2.0
django-webpush==0.3.4
kafka-python==2.0.2
drf-nested-routers==0.94.1
django_cleanup==9.0.0
django-cors-headers==4.5.0
google-generativeai==0.7.2
json-logic-qubit==0.9.1
sentry-sdk[django]==2.45.0
django-storages==1.14.6
boto3==1.42.9
aiohttp
websockets